RandomClue Optimization
Instead of randomly generating numbers and, more importantly, deleting
from and ArrayList, numbers are selected by iterating through an
iterator if and only if the number last given was incorrect. Thus there
is no need to perform the much more costly deletion in an ArrayList as
well as random number generation (if that even really matters too much).