Data Structures ListStack optimized equals method
The equals method was copied from ArrayStack and there, if the other object is also a collection of the same size, the memberwise comparison was done through a rever for loop and .get of the ArrayStack's members. Repeated .gets for a linked list is not good, thus an iterator is used instead.